Move entity extensions to patchwork-extensions-entity
