+++ patchwork-extensions-entity/src/main/java/net/minecraftforge/common/extensions/IForgeEntity.java
