+++ patchwork-extensions-entity/src/main/java/net/patchworkmc/mixin/extensions/entity/MixinEntityType.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.patchworkmc.mixin.extension;
+package net.patchworkmc.mixin.extensions.entity;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -27,11 +27,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import net.minecraft.entity.EntityType;
 
-import net.patchworkmc.impl.extension.PatchworkEntityTypeBuilderExtensions;
-import net.patchworkmc.impl.extension.PatchworkEntityTypeExtensions;
+import net.patchworkmc.impl.extensions.entity.PatchworkEntityTypeExtensions;
 
-@Mixin(EntityType.Builder.class)
-public class MixinEntityTypeBuilder implements PatchworkEntityTypeBuilderExtensions {
+@Mixin(EntityType.class)
+public class MixinEntityType implements PatchworkEntityTypeExtensions {
 	@Unique
 	private Integer updateInterval = null;
 	@Unique
@@ -39,38 +38,39 @@ public class MixinEntityTypeBuilder implements PatchworkEntityTypeBuilderExtensi
 	@Unique
 	private Boolean shouldRecieveVelocityUpdates = null;
 
-	@Inject(method = "build", at = @At("RETURN"))
-	private void onBuildReturn(String id, CallbackInfoReturnable<EntityType> cir) {
-		PatchworkEntityTypeExtensions type = (PatchworkEntityTypeExtensions) cir.getReturnValue();
-
-		if (updateInterval != null) {
-			type.patchwork$setUpdateInterval(updateInterval);
-		}
-
+	@Inject(method = "getMaxTrackDistance", at = @At("HEAD"), cancellable = true)
+	private void hookGetMaxTrackingDistance(CallbackInfoReturnable<Integer> cir) {
 		if (trackingRange != null) {
-			type.patchwork$setTrackingRange(trackingRange);
+			cir.setReturnValue(trackingRange);
 		}
+	}
 
+	@Inject(method = "getTrackTickInterval", at = @At("HEAD"), cancellable = true)
+	private void hookGetTrackTickInterval(CallbackInfoReturnable<Integer> cir) {
+		if (updateInterval != null) {
+			cir.setReturnValue(updateInterval);
+		}
+	}
+
+	@Inject(method = "alwaysUpdateVelocity", at = @At("HEAD"), cancellable = true)
+	private void hookAlwaysUpdateVelocity(CallbackInfoReturnable<Boolean> cir) {
 		if (shouldRecieveVelocityUpdates != null) {
-			type.patchwork$setShouldReceiveVelocityUpdates(shouldRecieveVelocityUpdates);
+			cir.setReturnValue(shouldRecieveVelocityUpdates);
 		}
 	}
 
 	@Override
-	public EntityType.Builder setUpdateInterval(int interval) {
+	public void patchwork$setUpdateInterval(int interval) {
 		this.updateInterval = interval;
-		return (EntityType.Builder) (Object) this;
 	}
 
 	@Override
-	public EntityType.Builder setTrackingRange(int range) {
+	public void patchwork$setTrackingRange(int range) {
 		this.trackingRange = range;
-		return (EntityType.Builder) (Object) this;
 	}
 
 	@Override
-	public EntityType.Builder setShouldReceiveVelocityUpdates(boolean value) {
+	public void patchwork$setShouldReceiveVelocityUpdates(boolean value) {
 		this.shouldRecieveVelocityUpdates = value;
-		return (EntityType.Builder) (Object) this;
 	}
 }
