+++ patchwork-extensions-entity/src/main/java/net/patchworkmc/mixin/extensions/entity/MixinServerPlayerEntity.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.patchworkmc.mixin.extension;
+package net.patchworkmc.mixin.extensions.entity;
 
 import net.minecraftforge.common.extensions.IForgeEntity;
 import org.spongepowered.asm.mixin.Mixin;
