--- patchwork-extensions-entity/src/main/resources/patchwork-MobEntity.accesswidener
@@ -1,3 +1,0 @@
-accessWidener v1 named
-
-accessible field net/minecraft/entity/mob/MobEntity goalSelector Lnet/minecraft/entity/ai/goal/GoalSelector;
