+++ settings.gradle
@@ -23,11 +23,11 @@ include 'patchwork-events-input'
 //include 'patchwork-events-lifecycle'
 //include 'patchwork-events-rendering'
 //include 'patchwork-events-world'
-//include 'patchwork-extensions'
+include 'patchwork-extensions'
 //include 'patchwork-extensions-bakedmodel'
 //include 'patchwork-extensions-block'
 //include 'patchwork-extensions-blockentity'
-//include 'patchwork-extensions-entity'
+include 'patchwork-extensions-entity'
 //include 'patchwork-extensions-item'
 //include 'patchwork-extensions-shearing'
 //include 'patchwork-fake-players'
